+++ Tests/Unit/Stubs/Office365.psm1
@@ -9140,7 +9140,7 @@ function Add-PnPTenantTheme {
    param(
        [Parameter(Mandatory = $true)]
        [System.String]
-       $Name,
+       $Identity,
 
        [Parameter()]
        [System.Boolean]
@@ -9156,9 +9156,39 @@ function Add-PnPTenantTheme {
    )
 }
 
+function New-AuditConfigurationPolicy
+{
+    [CmdletBinding()]
+    param(
+        [Parameter()]
+        [System.String]
+        $Workload
+    )
+}
+
+function Get-AuditConfigurationPolicy
+{
+    [CmdletBinding()]
+    param(
+        [Parameter()]
+        [System.String]
+        $Identity
+    )
+}
+
+function Remove-AuditConfigurationPolicy
+{
+    [CmdletBinding()]
+    param(
+        [Parameter(Mandatory = $true)]
+        [System.String]
+        $Identity
+    )
+}
+
 function Get-PnPUserProfileProperty {
     [CmdletBinding()]
-   param(
+    param(
        [Parameter()]
        [System.String]
        $Account